% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnvSurv-package.R
\docType{package}
\name{wnvSurv-package}
\alias{wnvSurv}
\alias{wnvSurv-package}
\title{wnvSurv: West Nile Virus Surveillance Functions}
\description{
Shared functions for West Nile virus mosquito surveillance work in the Ebel Lab. Consolidates data ingest, cleaning, abundance/pooled infection rate/vector index calculations, and standard plots used across the wnv-ss_trap_hx_combiner and wnv-ss-wkly_report pipelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rtobiaskoch/wnv-ss_functions}
  \item Report bugs at \url{https://github.com/rtobiaskoch/wnv-ss_functions/issues}
}

}
\author{
\strong{Maintainer}: Toby Koch \email{r.tobiaskoch@gmail.com}

}
\keyword{internal}
