% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zone_lvls}
\alias{zone_lvls}
\title{Larimer County trapping zone levels}
\format{
A character vector of length 8.
}
\source{
\code{config_culex_combine.yml} and \code{fun_wnv_s_clean.R} defaults in
\code{wnv-ss_trap_hx_combiner}.
}
\usage{
zone_lvls
}
\description{
Canonical ordering of mosquito trapping zones used across the Ebel Lab
surveillance pipelines. NW/NE/SE/SW are Fort Collins quadrants; FC is a
Fort Collins composite group; LV = Loveland, BE = Berthoud, BC = Boulder.
Provided as the default for \code{zone_lvls}/\code{zone_lvl} arguments throughout the
package.
}
\keyword{datasets}
